backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.0
openai==1.54.3
biopython==1.83
pandas==2.1.4
openpyxl==3.1.2
requests==2.31.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
gunicorn==21.2.0
beautifulsoup4==4.12.2
# langchain==0.3.0
# langchain-openai==0.2.0
# langgraph==0.2.28